backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
websockets==12.0
aiohttp 